Rename files and test patch
